requirements.txt
pillow
numpy
pytest
huggingface_hub
diffusers
opencv-python
torch
accelerate
transformers
controlnet_aux
matplotlib